QMAPS-1754 pr fixes
