+++ src/libs/geolocation.js
@@ -1,17 +1,8 @@
-import { openAndWaitForClose } from 'src/modals/GeolocationModal';
 import { fire } from 'src/libs/customEvents';
-
-const geolocationPermissions = {
-  PROMPT: 'prompt',
-  GRANTED: 'granted',
-  DENIED: 'denied',
-};
-
-let hasPermissionModalOpenedOnce = false;
 
 // Show custom Modal when opening the Directions panel for the first time
 // and if the browser's geolocation permission hasn't been granted yet
-export async function directionModalOnFirstVisit() {
+export async function isLocationAvailable() {
 
   // Some browsers (Safari, etc) do not implement Permissions API
   if (!window.navigator.permissions) {
@@ -20,13 +11,7 @@ export async function directionModalOnFirstVisit() {
 
   // granted or denied
   const p = await window.navigator.permissions.query({ name: 'geolocation' });
-  if (p.state !== geolocationPermissions.PROMPT) {
-    return p.state === geolocationPermissions.GRANTED;
-  }
-
-  if (hasPermissionModalOpenedOnce === true) {return;}
-  hasPermissionModalOpenedOnce = true;
-  return await openAndWaitForClose();
+  return p.state;
 }
 
 export function handleError(error) {
